+++ .idea/kotlinc.xml
@@ -4,6 +4,6 @@
     <option name="jvmTarget" value="1.8" />
   </component>
   <component name="KotlinJpsPluginSettings">
-    <option name="version" value="1.8.21" />
+    <option name="version" value="1.8.22" />
   </component>
 </project>
