--- insets/README.md
@@ -1,24 +1,0 @@
-# Insets for Jetpack Compose (Deprecated)
-
-[![Maven Central](https://img.shields.io/maven-central/v/com.google.accompanist/accompanist-insets)](https://search.maven.org/search?q=g:com.google.accompanist)
-
-> :warning: This library has been deprecated as official support is now available in Compose 1.2.0. Please see our [Migration Guide](https://google.github.io/accompanist/insets/) for how to migrate.
-
-For more information, visit the documentation: https://google.github.io/accompanist/insets
-
-## Download
-
-```groovy
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation "com.google.accompanist:accompanist-insets:<version>"
-}
-```
-
-Snapshots of the development version are available in [Sonatype's `snapshots` repository][snap]. These are updated on every commit.
-
-
-  [snap]: https://oss.sonatype.org/content/repositories/snapshots/com/google/accompanist/accompanist-insets/
