--- insets/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=accompanist-insets
-POM_NAME=Accompanist Insets library
-POM_PACKAGING=aar
