--- insets/src/test/resources/robolectric.properties
@@ -1,3 +1,0 @@
-# Pin SDK to 30 since Robolectric does not currently support API 31:
-# https://github.com/robolectric/robolectric/issues/6635
-sdk=30
