--- sample/src/main/java/com/google/accompanist/sample/insets/EdgeToEdgeLazyColumn.kt
@@ -1,124 +1,0 @@
-/*
- * Copyright 2020 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:Suppress("DEPRECATION")
-
-package com.google.accompanist.sample.insets
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.FloatingActionButton
-import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Face
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.core.view.WindowCompat
-import com.google.accompanist.insets.LocalWindowInsets
-import com.google.accompanist.insets.ProvideWindowInsets
-import com.google.accompanist.insets.navigationBarsHeight
-import com.google.accompanist.insets.rememberInsetsPaddingValues
-import com.google.accompanist.insets.ui.Scaffold
-import com.google.accompanist.insets.ui.TopAppBar
-import com.google.accompanist.sample.AccompanistSampleTheme
-import com.google.accompanist.sample.R
-import com.google.accompanist.sample.randomSampleImageUrl
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
-
-class EdgeToEdgeLazyColumn : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        // Turn off the decor fitting system windows, which means we need to through handling
-        // insets
-        WindowCompat.setDecorFitsSystemWindows(window, false)
-
-        setContent {
-            // Update the system bars to be translucent
-            val systemUiController = rememberSystemUiController()
-            val useDarkIcons = MaterialTheme.colors.isLight
-            SideEffect {
-                systemUiController.setSystemBarsColor(Color.Transparent, darkIcons = useDarkIcons)
-            }
-
-            AccompanistSampleTheme {
-                ProvideWindowInsets {
-                    Sample()
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun Sample() {
-    Scaffold(
-        topBar = {
-            // We use TopAppBar from accompanist-insets-ui which allows us to provide
-            // content padding matching the system bars insets.
-            TopAppBar(
-                title = { Text(stringResource(R.string.insets_title_list)) },
-                backgroundColor = MaterialTheme.colors.surface.copy(alpha = 0.95f),
-                contentPadding = rememberInsetsPaddingValues(
-                    LocalWindowInsets.current.statusBars,
-                    applyBottom = false,
-                ),
-                modifier = Modifier.fillMaxWidth()
-            )
-        },
-        bottomBar = {
-            // We add a spacer as a bottom bar, which is the same height as
-            // the navigation bar
-            Spacer(Modifier.navigationBarsHeight().fillMaxWidth())
-        },
-        floatingActionButton = {
-            FloatingActionButton(
-                onClick = { /* TODO */ }
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Face,
-                    contentDescription = "Face icon"
-                )
-            }
-        }
-    ) { contentPadding ->
-        Box {
-            // We apply the contentPadding passed to us from the Scaffold
-            LazyColumn(
-                contentPadding = contentPadding,
-                modifier = Modifier.fillMaxSize(),
-            ) {
-                items(items = listItems) { imageUrl ->
-                    ListItem(imageUrl, Modifier.fillMaxWidth())
-                }
-            }
-        }
-    }
-}
-
-private val listItems = List(40) { randomSampleImageUrl(it) }
