--- sample/src/main/java/com/google/accompanist/sample/insets/EdgeToEdgeLazyColumnWithBottomNav.kt
@@ -1,158 +1,0 @@
-/*
- * Copyright 2020 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:Suppress("DEPRECATION")
-
-package com.google.accompanist.sample.insets
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.BottomNavigationItem
-import androidx.compose.material.FabPosition
-import androidx.compose.material.FloatingActionButton
-import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Alarm
-import androidx.compose.material.icons.filled.CardTravel
-import androidx.compose.material.icons.filled.Face
-import androidx.compose.material.icons.filled.Fastfood
-import androidx.compose.material.icons.filled.HolidayVillage
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.core.view.WindowCompat
-import com.google.accompanist.insets.LocalWindowInsets
-import com.google.accompanist.insets.ProvideWindowInsets
-import com.google.accompanist.insets.rememberInsetsPaddingValues
-import com.google.accompanist.insets.ui.BottomNavigation
-import com.google.accompanist.insets.ui.Scaffold
-import com.google.accompanist.insets.ui.TopAppBar
-import com.google.accompanist.sample.AccompanistSampleTheme
-import com.google.accompanist.sample.R
-import com.google.accompanist.sample.randomSampleImageUrl
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
-
-class EdgeToEdgeLazyColumnWithBottomNav : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        // Turn off the decor fitting system windows, which means we need to through handling
-        // insets
-        WindowCompat.setDecorFitsSystemWindows(window, false)
-
-        setContent {
-            // Update the system bars to be translucent
-            val systemUiController = rememberSystemUiController()
-            val useDarkIcons = MaterialTheme.colors.isLight
-            SideEffect {
-                systemUiController.setSystemBarsColor(Color.Transparent, darkIcons = useDarkIcons)
-            }
-
-            AccompanistSampleTheme {
-                ProvideWindowInsets {
-                    Sample()
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun Sample() {
-    Scaffold(
-        topBar = {
-            // We use TopAppBar from accompanist-insets-ui which allows us to provide
-            // content padding matching the system bars insets.
-            TopAppBar(
-                title = { Text(stringResource(R.string.insets_title_list_bottomnav)) },
-                backgroundColor = MaterialTheme.colors.surface.copy(alpha = 0.95f),
-                contentPadding = rememberInsetsPaddingValues(
-                    LocalWindowInsets.current.statusBars,
-                    applyBottom = false,
-                ),
-            )
-        },
-        bottomBar = {
-            var selected by remember { mutableStateOf(0) }
-
-            BottomNavigation(
-                backgroundColor = MaterialTheme.colors.surface.copy(alpha = 0.95f),
-                contentPadding = rememberInsetsPaddingValues(
-                    LocalWindowInsets.current.navigationBars
-                )
-            ) {
-                BottomNavigationItem(
-                    selected = selected == 0,
-                    onClick = { selected = 0 },
-                    icon = { Icon(Icons.Default.Fastfood, contentDescription = null) }
-                )
-                BottomNavigationItem(
-                    selected = selected == 1,
-                    onClick = { selected = 1 },
-                    icon = { Icon(Icons.Default.CardTravel, contentDescription = null) }
-                )
-                BottomNavigationItem(
-                    selected = selected == 2,
-                    onClick = { selected = 2 },
-                    icon = { Icon(Icons.Default.HolidayVillage, contentDescription = null) }
-                )
-                BottomNavigationItem(
-                    selected = selected == 3,
-                    onClick = { selected = 3 },
-                    icon = { Icon(Icons.Default.Alarm, contentDescription = null) }
-                )
-            }
-        },
-        floatingActionButton = {
-            FloatingActionButton(onClick = { /* TODO */ }) {
-                Icon(
-                    imageVector = Icons.Default.Face,
-                    contentDescription = "Face icon"
-                )
-            }
-        },
-        floatingActionButtonPosition = FabPosition.Center,
-        isFloatingActionButtonDocked = true,
-    ) { contentPadding ->
-        Box {
-            // We apply the contentPadding passed to us from the Scaffold
-            LazyColumn(
-                contentPadding = contentPadding,
-                modifier = Modifier.fillMaxSize(),
-            ) {
-                items(items = listItems) { imageUrl ->
-                    ListItem(imageUrl, Modifier.fillMaxWidth())
-                }
-            }
-        }
-    }
-}
-
-private val listItems = List(40) { randomSampleImageUrl(it) }
